Refactor GlobalDrawer to remount on key changes instead of useEffect by separating form from drawer
