+++ apps/web/src/providers/GlobalDrawerProvider.tsx
@@ -1,34 +1,15 @@
-import { FC, ReactNode, createContext, useEffect, useState } from "react";
-import { useForm, Controller, SubmitHandler } from "react-hook-form";
-import {
-  ItemSchema,
-  type ItemFormInput,
-  type ItemFormOutput,
-} from "@app/shared";
-import { zodResolver } from "@hookform/resolvers/zod";
+import { FC, ReactNode, createContext, useMemo, useState } from "react";
 
-import {
-  DEFAULT_GLOBAL_DRAWER_PROPS,
-  DEFAULT_ITEM,
-  MAX_DESCRIPTION,
-} from "../constants/drawer";
+import ItemForm from "../components/ItemForm";
+import { DEFAULT_GLOBAL_DRAWER_PROPS } from "../constants/drawer";
 import HideIcon from "../components/Icons/HideIcon";
-import { Button } from "../components/ui/button";
-import { Checkbox } from "../components/ui/checkbox";
 import {
   Drawer,
-  DrawerClose,
   DrawerContent,
-  DrawerDescription,
-  DrawerFooter,
   DrawerHeader,
   DrawerTitle,
 } from "../components/ui/drawer";
-import { Input } from "../components/ui/input";
-import { Select } from "../components/ui/select";
-import { Textarea } from "../components/ui/textarea";
 import { GlobalDrawerContextType, GlobalDrawerProps } from "../types/drawer";
-import { cn } from "../utils";
 
 export const GlobalDrawerContext =
   createContext<GlobalDrawerContextType | null>(null);
@@ -52,59 +33,11 @@ const GlobalDrawerProvider: FC<GlobalDrawerProviderProps> = ({ children }) => {
     onConfirm,
   } = drawerProps;
 
-  const isUpdateMode = type === "update";
-
-  const {
-    register,
-    control,
-    handleSubmit,
-    formState: { errors, isSubmitting, isValid, isDirty },
-    watch,
-    reset,
-  } = useForm<ItemFormInput, any, ItemFormOutput>({
-    resolver: zodResolver(ItemSchema),
-    mode: "onChange",
-    defaultValues,
-  });
-
-  const descriptionValue = watch("description") ?? "";
-  const quantityValue = watch("quantity") as number | undefined;
-
-  useEffect(() => {
-    if (type === undefined) return;
-
-    if (type === "update" || type === "create") {
-      reset(defaultValues);
-    }
-  }, [defaultValues, reset, type]);
-
-  const clearForm = () => {
-    if (isUpdateMode) {
-      return reset({
-        ...defaultValues,
-        description: defaultValues.description ?? ""
-      });
-    }
-    return reset(DEFAULT_ITEM);
-  };
-
-  useEffect(() => {
-    if (!open && isDirty) {
-      clearForm();
-    }
-  }, [open, isDirty]);
-
-  const onSubmit: SubmitHandler<ItemFormOutput> = async (data) => {
-    // API Call
-    clearForm();
-    setOpen(false);
-    onConfirm &&
-      onConfirm({
-        ...data,
-        quantity: data.quantity ?? 1,
-        purchased: data.purchased ?? false,
-      });
-  };
+  const formKey = useMemo(() => {
+    const idPart = defaultValues?.id ?? "newId";
+    const modePart = type ?? "none";
+    return `${modePart}:${idPart}`;
+  }, [type, defaultValues?.id]);
 
   const openDrawer = (props: GlobalDrawerProps) => {
     setDrawerProps({ ...props });
@@ -125,7 +58,7 @@ const GlobalDrawerProvider: FC<GlobalDrawerProviderProps> = ({ children }) => {
             (e.currentTarget as HTMLElement).focus();
           }}
         >
-          <div className="h-full bg-white">
+          <div className="flex flex-col h-full bg-white">
             <DrawerHeader className="bg-drawerHeaderBg w-full h-16 pl-[30px] border-b-[0.5px] border-drawerBorderGray">
               <div className="w-full h-full flex items-center justify-between">
                 <DrawerTitle className="text-secondaryFont font-dosis">
@@ -141,157 +74,18 @@ const GlobalDrawerProvider: FC<GlobalDrawerProviderProps> = ({ children }) => {
                 </button>
               </div>
             </DrawerHeader>
-            <div className="pl-[30px] pt-7 pr-6">
-              <div className="flex flex-col font-nunito font-normal mb-4">
-                <h2 className="text-lg leading-6 text-primaryFont">
-                  Add an Item
-                </h2>
-                <div className="flex items-center justify-between">
-                  <DrawerDescription
-                    className={cn(
-                      "text-base leading-[22px] text-secondaryFont mt-2",
-                      descriptionTextClassName
-                    )}
-                  >
-                    {description}
-                  </DrawerDescription>
-                  {isDirty && (
-                    <button
-                      onClick={clearForm}
-                      className="font-medium text-xs hover:opacity-80 text-primaryFont"
-                    >
-                      Clear Form
-                    </button>
-                  )}
-                </div>
-              </div>
-              <form
-                id="item-form"
-                className="flex flex-col gap-4"
-                onSubmit={handleSubmit(onSubmit)}
-              >
-                <div>
-                  <Input
-                    id="item-name"
-                    placeholder="Item Name"
-                    {...register("itemName")}
-                  />
-                  {errors.itemName && (
-                    <p className="mt-1 text-sm text-red-600">
-                      {errors.itemName.message}
-                    </p>
-                  )}
-                </div>
-                <div>
-                  <div className="relative h-[140px]">
-                    <Textarea
-                      id="item-description"
-                      placeholder="Description"
-                      {...register("description")}
-                    />
-                    <span
-                      className={cn(
-                        "text-xs absolute bottom-3 right-3",
-                        descriptionValue.length > MAX_DESCRIPTION
-                          ? "text-red-600"
-                          : "text-secondaryFont"
-                      )}
-                      aria-live="polite"
-                    >
-                      {descriptionValue.length}/{MAX_DESCRIPTION}
-                    </span>
-                  </div>
-                  {errors.description && (
-                    <p className="mt-1 text-sm text-red-600">
-                      {errors.description.message}
-                    </p>
-                  )}
-                </div>
-                <div>
-                  <Controller
-                    control={control}
-                    name="quantity"
-                    render={({ field }) => (
-                      <Select
-                        open={selectOpen}
-                        onOpenChange={setSelectOpen}
-                        onClick={(value: number) => {
-                          field.onChange(value);
-                          setSelectOpen(false);
-                        }}
-                        value={quantityValue}
-                      />
-                    )}
-                  />
-                  {errors.quantity && (
-                    <p className="mt-1 text-sm text-red-600">
-                      {errors.quantity.message}
-                    </p>
-                  )}
-                </div>
-                {isUpdateMode && (
-                  <div className="flex items-center justify-between mt-4 z-[1]">
-                    <Controller
-                      control={control}
-                      name="purchased"
-                      render={({ field: { value, onChange, onBlur, ref } }) => (
-                        <div className="flex gap-4 h-6 items-center">
-                          <Checkbox
-                            id="item-purchased"
-                            name="item-purchased"
-                            checked={value}
-                            onCheckedChange={(checked) =>
-                              onChange(Boolean(checked))
-                            }
-                            aria-invalid={!!errors.purchased || undefined}
-                          />
-                          <label
-                            htmlFor="item-purchased"
-                            className={cn(
-                              "text-placeholderGray text-base leading-[22px] font-nunito",
-                              value && "text-primaryFont"
-                            )}
-                          >
-                            Purchased
-                          </label>
-                        </div>
-                      )}
-                    />
-                  </div>
-                )}
-              </form>
-            </div>
+            <ItemForm
+              key={formKey}
+              type={type}
+              description={description}
+              descriptionTextClassName={descriptionTextClassName}
+              defaultValues={defaultValues}
+              onConfirm={onConfirm}
+              onClose={closeDrawer}
+              selectOpen={selectOpen}
+              setSelectOpen={setSelectOpen}
+            />
           </div>
-          <DrawerFooter className="bg-white w-full flex">
-            <div className="flex gap-4 w-full justify-end pr-2 font-nunito mb-1">
-              <DrawerClose asChild>
-                <Button
-                  variant="secondary"
-                  className="h-9 font-normal hover:opacity-80"
-                >
-                  Cancel
-                </Button>
-              </DrawerClose>
-              <Button
-                variant="default"
-                className={cn(
-                  "h-9 font-normal w-[85px]",
-                  isUpdateMode && "w-[100px]"
-                )}
-                form="item-form"
-                type="submit"
-                disabled={isSubmitting || !isValid || (!isDirty && isUpdateMode)}
-              >
-                {isSubmitting
-                  ? isUpdateMode
-                    ? "Updating..."
-                    : "Saving..."
-                  : isUpdateMode
-                  ? "Save Item"
-                  : "Add Task"}
-              </Button>
-            </div>
-          </DrawerFooter>
         </DrawerContent>
       </Drawer>
     </GlobalDrawerContext.Provider>
